+++ src/main/java/by/triumgroup/recourse/repository/CourseFeedbackRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
-public interface CourseFeedbackRepository extends PagingAndSortingRepository<CourseFeedback, Long> {
+public interface CourseFeedbackRepository extends PagingAndSortingRepository<CourseFeedback, Integer> {
 
-    List<CourseFeedback> findByCourseIdOrderByIdDesc(Long courseId);
+    List<CourseFeedback> findByCourseIdOrderByIdDesc(Integer courseId);
 
-    List<CourseFeedback> findByCourseIdOrderByIdDesc(Long courseId, Pageable pageable);
+    List<CourseFeedback> findByCourseIdOrderByIdDesc(Integer courseId, Pageable pageable);
 
 }
