+++ src/main/java/by/triumgroup/recourse/repository/CourseRepository.java
@@ -8,17 +8,17 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CourseRepository extends PagingAndSortingRepository<Course, Long> {
+public interface CourseRepository extends PagingAndSortingRepository<Course, Integer> {
 
     List<Course> findByTitleContainingIgnoreCaseOrderByIdDesc(String title, Pageable pageable);
 
-    List<Course> findByTeacherIdOrderByIdDesc(Long id, Pageable pageable);
+    List<Course> findByTeacherIdOrderByIdDesc(Integer id, Pageable pageable);
 
-    List<Course> findByTeacherIdAndStatusOrderByIdDesc(Long id, Course.Status status, Pageable pageable);
+    List<Course> findByTeacherIdAndStatusOrderByIdDesc(Integer id, Course.Status status, Pageable pageable);
 
-    List<Course> findByOrganizerIdOrderByIdDesc(Long id, Pageable pageable);
+    List<Course> findByOrganizerIdOrderByIdDesc(Integer id, Pageable pageable);
 
-    List<Course> findByOrganizerIdAndStatusOrderByIdDesc(Long id, Course.Status status, Pageable pageable);
+    List<Course> findByOrganizerIdAndStatusOrderByIdDesc(Integer id, Course.Status status, Pageable pageable);
 
     List<Course> findByStatusOrderByIdDesc(Course.Status status, Pageable pageable);
 }
