+++ src/main/java/by/triumgroup/recourse/repository/HometaskSolutionRepository.java
@@ -6,14 +6,14 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.util.List;
 
-public interface HometaskSolutionRepository extends PagingAndSortingRepository<HometaskSolution, Long> {
+public interface HometaskSolutionRepository extends PagingAndSortingRepository<HometaskSolution, Integer> {
 
-    List<HometaskSolution> findByHometaskId(Long id);
+    List<HometaskSolution> findByHometaskId(Integer id);
 
-    List<HometaskSolution> findByHometaskId(Long id, Pageable pageable);
+    List<HometaskSolution> findByHometaskId(Integer id, Pageable pageable);
 
-    List<HometaskSolution> findByStudentId(Long id);
+    List<HometaskSolution> findByStudentId(Integer id);
 
-    List<HometaskSolution> findByStudentId(Long id, Pageable pageable);
+    List<HometaskSolution> findByStudentId(Integer id, Pageable pageable);
 
 }
