+++ src/main/java/by/triumgroup/recourse/repository/UserRepository.java
@@ -5,8 +5,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends PagingAndSortingRepository<User, Long> {
-
+public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
     User findByEmail(String email);
 
 }
