+++ src/main/java/by/triumgroup/recourse/service/UserService.java
@@ -3,7 +3,7 @@ package by.triumgroup.recourse.service;
 import by.triumgroup.recourse.entity.User;
 import by.triumgroup.recourse.service.exception.ServiceException;
 
-public interface UserService extends CrudService<User, Long> {
+public interface UserService extends CrudService<User, Integer> {
 
     User findByEmail(String email) throws ServiceException;
